Introduce BaseIndex and hoist query-agnostic retrieval primitives
Add a thin BaseIndex (close/mem_usage/type + IndexType category) so a
heterogeneous index cache can hold vector and future full-text indexes
behind one handle. Hoist the query-agnostic primitives out of the vector
header into retrieval top-level: RowIdFilter/BitmapRowIdFilter into
row_id_filter.h, and the scored output and iterator into scored_result.h,
renaming AnnIterator to ScoredRowIterator and VectorAnnResult to
ScoredResult. VectorAnnIndex now derives from BaseIndex, reports
IndexType::VECTOR, and exposes the ANN algorithm via algo_type().
